Rename UI to Shared and lowercase components folders
