+++ src/components/detailtabs/styled.js
