+++ src/components/shared/PasteButton.js
