+++ src/components/shared/styled.js
