+++ src/screens/TransactionDetailScreen.js
@@ -13,9 +13,9 @@ import base64js from 'base64-js'
 import crypto from 'crypto-js'
 import sha256 from 'crypto-js/sha256';
 import { get, sortBy } from 'lodash'
-import ErrorMessage from './../components/UI/ErrorMessage'
-import SecurityForm from './../components/UI/SecurityForm'
-import DetailTabs from './../components/DetailTab/DetailTabs'
+import ErrorMessage from './../components/shared/ErrorMessage'
+import SecurityForm from './../components/shared/SecurityForm'
+import DetailTabs from './../components/detailtabs/DetailTabs'
 import { decodeFromXdr, signXdr } from './../utils/xdrUtils';
 import { generateKeypair } from './../utils/bipUtil';
 import PouchDB from 'pouchdb-react-native'
